POC for enabling model optimization via CompileAPI for non-compiling EPs as a separate mode.
