+++ onnxruntime/core/session/model_compilation_options.h
@@ -234,7 +234,8 @@ class ModelCompilationOptions {
   std::filesystem::path input_model_path_;
   const void* input_model_data_ = nullptr;
   size_t input_model_data_size_ = 0;
-  const OrtModel* input_model_ = nullptr;  // Borrowed pointer
+  const OrtModel* input_model_ = nullptr;     // Borrowed pointer
+  bool user_set_optimization_level_ = false;  // true if caller explicitly called SetGraphOptimizationLevel
 };
 }  // namespace onnxruntime
 #endif  // !defined(ORT_MINIMAL_BUILD)
